Population Component Duplication Refactor
