+++ src/state-judicial-district-map/index.js
@@ -1,0 +1,1 @@
+export { default } from "./StateJudicialDistrictMap";
