--- src/state-office-map/index.js
@@ -1,1 +1,0 @@
-export { default } from "./StateOfficeMap";
